ES: AIS: Ais server not launching on boot from earlyservices
V4L subdev node creation was delayed as it was done part of
lateinit_call so nodes were not available when server was trying to
open. Create a new driver(cam_v4l_nodes) which will create v4l subdev nodes
in module init. Looping for the available nodes not required now.

Change-Id: Ie3834748a50d2415816546b4acabb2fc95ba9f6b