earlyservices : create all drm cards required for es
As there are two DPUs that will be probed in gen4, the
drm card numbers can sometimes vary. We are seeing this
happens approx 20% of the time.

In the early app launch path, the respective cards number
and app are tightly coupled. Since there is a possibilty
for mismatch with the variation in card numbers for
shared display, the existing card creation logic might
cause the client app to fail as it will not be opening
the correct card.

If more shared display card are introduced going forward,
this variation will create more risk for mismatch.

So, when the display modules are loaded we create all the
cards in the `dev/dri` path that are used by the early apps.
This ensures that the cards are available for early apps
(esplash and ervc) before the apps are launched.

Change-Id: I6cbbfc00f9b9a795e5321de57430592570861139
Signed-off-by: Yashovardhan Bhatt <quic_ybhatt@quicinc.com>