early_service: add card2 in the creation list
DRI card number assignment is not static and it can change depending on the
order in which the respective drm device are registered and have the minor
numbers assigned.

In some of the cases the card number assignment gets changed. current
implementation in early service code is only creating specific card3, card4
and card5.

This change add the card2 to the list and let the DRM Client
decide which card it requires to open by matching against the name of the card.

Change-Id: I8dce05f367c14627e58943e25467a2fbf2c9a96a
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>
(cherry picked from commit 191410ccfafb01e9a64098ee2627577692fcc38b)