ES : Fix for bootslot not found.
As boot slot is not part of cmdline FW mount is failing
thereby causing es apps launch failure which depend on FW.
Some of the config keys are not available in kernel
cmdline but are available in bootconfig.
Irrespective of API32 level, check for both kernel cmdline
and bootconfig for the config keys.

Change-Id: I45e0b49e828ac51124e1d533ba4610ceac7acaca
Signed-off-by: Sunish Nambayil <quic_snambayi@quicinc.com>