early_init: synchronous system booting during early service execution
After waiting for storage and display to complete,
it sends the "kedone" message to notify the init process
that it can continue system booting synchronously.
Advances the time to create the linker64 connection.
The "sedone" message is still sent after the early service completes.
Add a delay and fix the function wait_for_early_apps.

Change-Id: Ie07d8b45ad29a1a4b8cfbad0ca93a2fd7f04c76d
Signed-off-by: Yinfei Wang <yinfwang@qti.qualcomm.com>